+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `isManual` option to `useInfiniteScrollDataLoad()`
+
 ## [2.4.0] - 2022-10-19
 
 ### Changed