+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `isManual` option to `useInfiniteScrollDataLoad()`
+- Disabled validation for read only forms
 
 ## [2.4.0] - 2022-10-19