+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.0] - 2022-11-24
+
 ### Changed
 
 - Form Store table to load results from a button instead of scrolling to the bottom of the window